ENH: Updates from tbb44_20150928oss_src.tgz
Intel TBB 4.4
TBB_INTERFACE_VERSION == 9000

Provide updates from tarball.

+++ src/test/test_task_scheduler_init.cpp
@@ -48,7 +48,7 @@ size_t __TBB_EXPORTED_FUNC get_initial_auto_partitioner_divisor();
 }}
 
 int ArenaConcurrency() {
-    return tbb::internal::get_initial_auto_partitioner_divisor()/4; // TODO: expose through task_arena interface?
+    return int(tbb::internal::get_initial_auto_partitioner_divisor()/4); // TODO: expose through task_arena interface?
 }
 
 // Generally, TBB does not guarantee mandatory parallelism. This test uses some whitebox knowledge about when all the threads can be available
